requirements.txt
asgiref==3.7.2
cfgv==3.3.1
distlib==0.3.6
Django==3.2
django-debug-toolbar==3.4.0
django-split-settings==1.2.0
Faker==18.11.1
filelock==3.12.2
flake8==4.0.0
identify==2.5.24
iniconfig==2.0.0
mccabe==0.6.1
nodeenv==1.8.0
packaging==23.1
platformdirs==3.8.0
pluggy==1.2.0
pre-commit==3.3.3
psycopg2==2.9.6
psycopg2-binary==2.9
pycodestyle==2.8.0
pyflakes==2.4.0
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
virtualenv==20.23.1
